Fix pointer type cast errors from Visual Studio
Visual Studio reports more pointer type cast errors with 64-bit build.
This will cover the issue on the existing targets.

+++ BootloaderCommonPkg/Library/FatLib/FatLib.c
@@ -71,7 +71,7 @@ GetNextFilePathNode (
   if (Ptr != NULL) {
     while (TRUE) {
       if ((*Ptr == 0) || (*Ptr == '\\') || (*Ptr == '/')) {
-        Len = Ptr - FileName;
+        Len = (UINT32)(Ptr - FileName);
         break;
       }
       Ptr++;
